+++ src/gps_tracker/client/__init__.py
@@ -1,0 +1,6 @@
+"""Definitions of subpackage gps_tracker.client."""
+
+from gps_tracker.client.asynchronous import AsyncClient
+from gps_tracker.client.config import Config
+from gps_tracker.client.datatypes import Device, Tracker, TrackerData
+from gps_tracker.client.synchronous import Client