+++ src/invoxia/client/__init__.py
@@ -1,6 +1,0 @@
-"""Definitions of subpackage invoxia.client."""
-
-from invoxia.client.asynchronous import AsyncClient
-from invoxia.client.config import Config
-from invoxia.client.datatypes import Device, Tracker, TrackerData
-from invoxia.client.synchronous import Client