+++ tests/test_invoxia/__init__.py
@@ -1,1 +1,0 @@
-"""Unit tests for `invoxia` package."""