+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Changelog
 0.1.0
 -----
 
-- Implement the synchronous :class:`Client <invoxia.client.sync.Client>`
+- Implement the synchronous :class:`Client <gps_tracker.client.sync.Client>`
 - Document the use of :doc:`current module <api/modules>` and :doc:`quickstart <start>`
-- Add :mod:`enumerations <invoxia.client.datatypes>` to improve readability of some tracker attributes.
+- Add :mod:`enumerations <gps_tracker.client.datatypes>` to improve readability
+  of some tracker attributes.